requirements.txt
Django==5.1.3
transformers==4.46.2
python-dotenv==1.0.1
google-api-python-client==2.153.0